Recombine input datasets

When aggregating data, the input datasets are split up to make them more manageable. The split datasets are stored in the 'data' folder, and the whole datasets are removed from the global environment, until aggregation is complete. If you need to manually restore the original input datasets to the global environment, you can do so using this function.

Usage

recombine_input_data(
  data_file_path,
  person_data_input_name = "person_data",
  vaccination_data_input_name = "vaccination_data",
  outcome_data_input_name = "outcome_data",
  patient_data_input_name = NULL
)

Arguments

data_file_path

Path to 'data' folder where split datasets are stored

person_data_input_name

Desired name of person_data object in global environment, defaults to "person_data". Set to NULL to skip this dataset.

vaccination_data_input_name

Desired name of vaccination_data object in global environment, defaults to "vaccination_data". Set to NULL to skip this dataset.

outcome_data_input_name

Desired name of outcome_data object in global environment, defaults to "outcome_data". Set to NULL to skip this dataset.

patient_data_input_name

Deprecated. Use person_data_input_name instead.

Examples

if (FALSE) { # \dontrun{
# Restore split input data back to the global environment
recombine_input_data(
  data_file_path = "path/to/results/data"
)
} # }