Aggregate line list datasets

This function will take the input datasets and create a folder containing the aggregated tables as separate csv files, depending on your participation level. A separate folder is created for each cycle period.

Aggregation can be stopped and resumed (see resume_cycle_aggregation() function).

To trigger historical data aggregation (i.e. before vaccinations began), set vaccination_data = NULL.

Requires an 'options.txt' file to run - this will be sent to you by GVDN.

Validation of the input datasets is automatically run as part of this function, and will throw an error if it finds any issues that require fixing. If you would like to run validation separately, use the validate_input_data() function.

Usage

aggregate_data(
  patient_data,
  vaccination_data,
  outcome_data,
  cycle_start_date,
  cycle_end_date,
  site_code,
  participation_level,
  options_file_location = getwd(),
  options = NULL,
  suppression_limit = NULL,
  split_size = 5e+05,
  stop_after_n_splits = NULL,
  restore_input_data = TRUE,
  working_directory = getwd(),
  input_data_in_chunks = FALSE,
  final_chunk = FALSE
)

Arguments

patient_data

Patient data input data.table

vaccination_data

Vaccination data input data.table. Set to NULL to perform historical data aggregation

outcome_data

AESI outcome data input data.table

cycle_start_date

Start date of cycle, as a date object, an 8 digit number of the form yyyymmdd, or a string of the form yyyy-mm-dd

cycle_end_date

End date of cycle, as a date object, an 8 digit number of the form yyyymmdd, or a string of the form yyyy-mm-dd

site_code

Site code (e.g. AUS_MCRI), as a string

participation_level

Site's level of participation, as a number from 1-3: 1 = minimal, 2 = partial, 3 = full

options_file_location

Location where options.txt file is stored. Supply either options_file_location or options, not both. Defaults to getwd()

options

Options object, already converted from txt file. Supply either options_file_location or options, not both

suppression_limit

If required, low numbers can be suppressed. Numbers less than the limit will be suppressed

split_size

Input data will be automatically split into pieces of this size (default 500,000). If memory is an issue, lower the split_size. Aggregation is done on the split data separately, and then combined once all splits are complete. Aggregation can be stopped and restarted from the last completed split (see resume_cycle_aggregation() function)

stop_after_n_splits

Use this to stop aggregation after a given number of splits. Aggregation can then be resumed at a later date (see resume_cycle_aggregation() function)

restore_input_data

If TRUE (default) the input data will be restored from the data folder to the global environment once the aggregation is complete, or stop_after_n_splits is reached. If FALSE, the input data will not be restored. WARNING: the data folder is removed once aggregation is complete (not if stopped using stop_after_n_splits)

working_directory

Location of working directory. This is where the folder for this cycle will be created. Defaults to getwd().

input_data_in_chunks

Option to pre-split input data. Only use this if your data is too large to load into R in one go (otherwise just use split_size)

final_chunk

If you have pre-split your input data, use this to indicate when you are running the last chunk of your input data. This will trigger the function to combine all the temporary results at the end